Вынес общий тестовый класс, обновил отчет аллюр, исправил не корректные тесты
