+++ src/test/java/ru/praktikum/OpenConstructorTest.java
@@ -1,29 +1,13 @@
 package ru.praktikum;
 
 import io.qameta.allure.junit4.DisplayName;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
-import org.openqa.selenium.WebDriver;
-import ru.praktikum.data.User;
-import ru.praktikum.data.UserGenerator;
-import ru.praktikum.extensions.WebDriverFactory;
 import ru.praktikum.pages.MainPage;
-import ru.praktikum.user.UserAction;
 
 import static org.junit.Assert.assertTrue;
 
-public class OpenConstructorTest {
-    private WebDriver driver;
-    private User user;
-    private UserAction userAction;
-    @Before
-    public void init() {
-        driver = WebDriverFactory.get();
-        userAction = new UserAction();
-        user = UserGenerator.getUser();
-        userAction.createUser(user);
-    }
+public class OpenConstructorTest extends BaseTestWithCreateUser{
+
     @Test
     @DisplayName("Open constructor from constructor button")
     public void openConstructorFromConstructorBtnTest(){
@@ -35,6 +19,7 @@ public class OpenConstructorTest {
                 .isVisibleCreateOrderBtn();
         assertTrue("Ошибка при переходе  из личного кабинета в конструктор по клику на «Конструктор»", isVisibleCreateOrderBtn);
     }
+
     @Test
     @DisplayName("Open constructor from logo button")
     public void openConstructorFromLogoTest(){
@@ -47,9 +32,4 @@ public class OpenConstructorTest {
         assertTrue("Ошибка при переходе  из личного кабинета в конструктор по клику на логотип", isVisibleCreateOrderBtn);
     }
 
-    @After
-    public void tearDown() {
-        driver.quit();
-        userAction.deleteUser(userAction.login(user));
-    }
 }
