+++ src/test/java/ru/praktikum/OpenIngridientsSectionTest.java
@@ -1,28 +1,13 @@
 package ru.praktikum;
 
 import io.qameta.allure.junit4.DisplayName;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
-import org.openqa.selenium.WebDriver;
-import ru.praktikum.data.User;
-import ru.praktikum.data.UserGenerator;
-import ru.praktikum.extensions.WebDriverFactory;
 import ru.praktikum.pages.MainPage;
-import ru.praktikum.user.UserAction;
 
 import static org.junit.Assert.assertTrue;
 
-public class OpenIngridientsSectionTest {
-    private WebDriver driver;
-    private User user;
-    private UserAction userAction;
-    @Before
-    public void init() {
-        driver = WebDriverFactory.get();
-        userAction = new UserAction();
-        user = UserGenerator.getUser();
-    }
+public class OpenIngridientsSectionTest extends BaseTestWithoutCreateUser{
+
     @Test
     @DisplayName("Open bun section in constructor")
     public void openBunSectionTest(){
@@ -32,6 +17,7 @@ public class OpenIngridientsSectionTest {
                 .isVisibleBunSection();
         assertTrue("Ошибка при переходе к разделу Булки", isVisibleBunSection);
     }
+
     @Test
     @DisplayName("Open souce section in constructor")
     public void openSouceSectionTest(){
@@ -40,6 +26,7 @@ public class OpenIngridientsSectionTest {
                 .isVisibleSouceSection();
         assertTrue("Ошибка при переходе к разделу Соусы", isVisibleSouceSection);
     }
+
     @Test
     @DisplayName("Open filling section in constructor")
     public void openFillingSectionTest(){
@@ -47,10 +34,5 @@ public class OpenIngridientsSectionTest {
                 .clickFillingSectionBtn()
                 .isVisibleFillingSection();
         assertTrue("Ошибка при переходе к разделу Начинки", isVisibleFillingSection);
-    }
-
-    @After
-    public void tearDown() {
-        driver.quit();
     }
 }
